+++ test/config-with-other-directory/.postgratorrc.cjs
@@ -1,0 +1,9 @@
+module.exports = {
+    migrationPattern: 'other-migrations/*',
+    driver: 'pg',
+    host: '127.0.0.1',
+    port: 5432,
+    database: 'postgrator',
+    username: 'postgrator',
+    password: 'postgrator',
+}