+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 5.0.0
+### Unreleased
+* (**Breaking**) Node.js 12 or later required
+* (**Breaking**) Removed option -t, --detect-version-conflicts. Conflicts are now always detected.
+* (**Breaking**) Target postgrator 5.x.x
+* (**Breaking**) Migrate to using cosmiconfig for configuration file
+* (**Breaking**) `secure` option is now `ssl` and it works differently for each supported driver
+* Add support for node 16.x.x
+* Support merging cli and config file options
+
 ## 4.0.0
 ### May, 17, 2020
 * (**Breaking**) Node.js 10 or later required